Physics components almost done. (needs further testing)

+++ server/inc/Components/Abstracts/AGraphicsComponent.hpp
@@ -16,9 +16,9 @@ namespace Component {
 		explicit AGraphicsComponent(Engine::Entity *parentEntity, RessourcesLoader *resourceLoader);
 		~AGraphicsComponent() override = default;
 
-		bool isInArea(float xPos, float yPos, float xRange, float yRange);
-		float getRelativeXPos(float xPos, float xRange);
-		float getRelativeYPos(float yPos, float yRange);
+		bool isInArea(double xPos, double yPos, double xRange, double yRange);
+		double getRelativeXPos(double xPos, double xRange);
+		double getRelativeYPos(double yPos, double yRange);
 
 		RessourcesLoader *getResourceLoader() const;
 		void setResourceLoader(RessourcesLoader *_resourceLoader);
